+++ src/parser/inline/comment.ts
@@ -1,8 +1,10 @@
 ﻿import { CommentParser } from '../inline';
-import { match } from '../../combinator';
+import { match, verify } from '../../combinator';
 import '../source/unescapable';
 import { html } from 'typed-dom';
 
-export const comment: CommentParser = match(
-  /^<(#+)\s+([\s\S]*?\s+)?\1>/,
-  ([, , s = ''], rest) => [[html('sup', { class: 'comment', title: s.trim() })], rest]);
+export const comment: CommentParser = verify(
+  match(
+    /^<(#+)\s+([\s\S]*?\s+)?\1>/,
+    ([, , s = ''], rest) => [[html('sup', { class: 'comment', title: s.trim() })], rest]),
+  ([el]) => el.title.trim() !== '');