+++ src/parser/api/source.ts
@@ -1,6 +1,7 @@
 import { BlockParser } from '../block';
 import { bind } from '../../combinator';
 import { parse } from './parse';
+import { concat } from 'spica/concat';
 
 const sources: WeakMap<Element, string> = new WeakMap();
 
@@ -8,7 +9,13 @@ export function breaklines(source: string): string {
   return [...parse(source).children]
     .map<string>(el => {
       if (el instanceof HTMLParagraphElement === false) return sources.get(el)!;
-      const breaks = el.querySelectorAll('br, .linebreak');
+      const breaks = [...el.querySelectorAll<HTMLElement>('br, .linebreak, .comment')]
+        .reduce<Element[]>((acc, el) =>
+          concat(acc,
+            el.matches('.comment')
+              ? Array<Element>(el.title.split('\n').length - 1).fill(el)
+              : [el])
+        , []);
       return sources.get(el)!
         .split('\n')
         .map((line, i) =>